module_trade_execution/requirement.txt
python-binance # 幣安 API 庫 
google-cloud-pubsub
google-cloud-secret-manager

pandas 
